+++ hydrant/cli.py
@@ -6,6 +6,7 @@ DEFAULT_AWS_REGION = 'us-east-1'
 
 
 @click.command(epilog='Source: https://github.com/bwbaugh/hydrant')
+@click.version_option()
 @click.argument('delivery_stream')
 @click.option(
     '--region',
@@ -14,8 +15,17 @@ DEFAULT_AWS_REGION = 'us-east-1'
     show_default=True,
     metavar='AWS_REGION',
 )
-@click.version_option()
-def main(delivery_stream, region):
+@click.option(
+    '--print-record-id/--no-print-record-id',
+    default=False,
+    help=(
+        'Controls outputting the ID of each record, which is a unique '
+        'string assigned to each record. Producer applications can '
+        'use this ID for purposes such as auditability and '
+        'investigation.'
+    ),
+)
+def main(delivery_stream, region, print_record_id):
     """Redirects stdin to Amazon Kinesis Firehose.
 
     Records will be written to DELIVERY_STREAM. Data should be
@@ -25,10 +35,12 @@ def main(delivery_stream, region):
     """
     client = _get_firehose_client(region_name=region)
     for line in click.get_binary_stream('stdin'):
-        client.put_record(
+        response = client.put_record(
             DeliveryStreamName=delivery_stream,
             Record={'Data': line},
         )
+        if print_record_id:
+            click.echo(message=response['RecordId'])
 
 
 def _get_firehose_client(region_name):